massively simplify implementation
Instead of opening each layer dir and parsing the parent field in the json file
we use the manifest.json file. It contains the layers in the correct order
(parent to youngest child).
This allows us to get rid of the list.h stuff and avoids the overhead of finding
and returning the correct child of each non-root layer.

Signed-off-by: Christian Brauner <christian.brauner@mailbox.org>